Incremental fix which creates shortcuts
But this branch will probably get abandoned now since the links created
will launch fine when the folder is navigated to and double clicked
on... but they (irritatingly) will not work when launched from a
toolbar. I'm guessing this is a signing / security thing but after
trying hex decoding and seeing where the issue is likely to be, it's way
above the effort level required here if its even possible.

The only workable alternative would be to generate a powershell script
which the user could then execute, but that in itself would be a bit of
a pain in the ass to support.

Plus with X11 GUI support coming for WSL, and the vsock workaround in
https://github.com/nbdd0121/x11-over-vsock, I'm over this as a priority
so instead I think I'll focus on usability. I'll leave this branch here
though in case I (or someone) wants to pick up the fight in future.